simplify and fix xpbd example
