+++ examples/xpbd.rs
@@ -1,0 +1,118 @@
+//! A simple example showing how to use oxidized_navigation with xpbd.
+
+use bevy::{prelude::*, DefaultPlugins};
+use bevy_xpbd_3d::prelude::{Collider, PhysicsPlugins};
+use oxidized_navigation::{
+    debug_draw::{DrawNavMesh, OxidizedNavigationDebugDrawPlugin},
+    NavMeshAffector, NavMeshSettings, OxidizedNavigationPlugin,
+};
+
+fn main() {
+    App::new()
+        // Default Plugins
+        .add_plugins((
+            DefaultPlugins,
+            OxidizedNavigationPlugin::<Collider>::new(NavMeshSettings {
+                cell_width: 0.25,
+                cell_height: 0.1,
+                tile_width: 100,
+                world_half_extents: 250.0,
+                world_bottom_bound: -100.0,
+                max_traversable_slope_radians: (40.0_f32 - 0.1).to_radians(),
+                walkable_height: 20,
+                walkable_radius: 1,
+                step_height: 3,
+                min_region_area: 100,
+                merge_region_area: 500,
+                max_contour_simplification_error: 1.1,
+                max_edge_length: 80,
+                max_tile_generation_tasks: Some(9),
+            }),
+            OxidizedNavigationDebugDrawPlugin,
+            PhysicsPlugins::default(),
+        ))
+        .add_systems(Startup, setup)
+        .add_systems(Update, (toggle_nav_mesh_system,))
+        .run();
+}
+
+//
+//  Toggle drawing Nav-mesh.
+//  Press M to toggle drawing the navmesh.
+//
+fn toggle_nav_mesh_system(keys: Res<Input<KeyCode>>, mut show_navmesh: ResMut<DrawNavMesh>) {
+    if keys.just_pressed(KeyCode::M) {
+        show_navmesh.0 = !show_navmesh.0;
+    }
+}
+
+fn setup(
+    mut commands: Commands,
+    mut meshes: ResMut<Assets<Mesh>>,
+    mut materials: ResMut<Assets<StandardMaterial>>,
+) {
+    // Plane
+    commands.spawn((
+        PbrBundle {
+            mesh: meshes.add(Mesh::from(bevy::prelude::shape::Plane {
+                size: 50.0,
+                subdivisions: 0,
+            })),
+            material: materials.add(Color::rgb(0.3, 0.5, 0.3).into()),
+            transform: Transform::IDENTITY,
+            ..default()
+        },
+        Collider::cuboid(50.0, 0.2, 50.0),
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+    ));
+
+    // Cube
+    commands.spawn((
+        PbrBundle {
+            mesh: meshes.add(Mesh::from(bevy::prelude::shape::Cube { size: 2.5 })),
+            material: materials.add(Color::rgb(0.1, 0.1, 0.5).into()),
+            transform: Transform::from_xyz(-5.0, 0.8, -5.0),
+            ..default()
+        },
+        Collider::cuboid(2.5, 2.5, 2.5),
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+    ));
+
+    // Tall Cube
+    commands.spawn((
+        PbrBundle {
+            mesh: meshes.add(Mesh::from(bevy::prelude::shape::Cube { size: 37.5 })),
+            material: materials.add(Color::rgb(0.1, 0.75, 0.5).into()),
+            transform: Transform::from_xyz(-0.179, 18.419, -27.744),
+            ..default()
+        },
+        Collider::cuboid(37.5, 37.5, 37.5),
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+    ));
+
+    // Thin wall
+    commands.spawn((
+        PbrBundle {
+            mesh: meshes.add(Mesh::from(bevy::prelude::shape::Cube { size: 0.1 })),
+            material: materials.add(Color::rgb(0.1, 0.1, 0.5).into()),
+            transform: Transform::from_xyz(-3.0, 0.8, 5.0).with_scale(Vec3::new(50.0, 15.0, 1.0)),
+            ..default()
+        },
+        // At the time of writing, xpbd (v0.2) colliders don't support scaling, so you have to create the collider with the post-scaled size.
+        Collider::cuboid(5.0, 1.5, 0.1),
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+    ));
+
+    // light
+    commands.spawn(PointLightBundle {
+        transform: Transform::from_xyz(4.0, 8.0, 4.0),
+        ..default()
+    });
+
+    // Camera
+    commands.spawn(Camera3dBundle {
+        transform: Transform::from_xyz(10.0, 10.0, 15.0)
+            .looking_at(Vec3::new(0.0, 2.0, 0.0), Vec3::Y),
+        ..default()
+    });
+}
